refactor: ViewとFormの責務を分割
twitterの予約投稿処理とX API呼び出しをservicesへ抽出し、eventのPDF/画像検証、community管理フォームの副作用、user_accountのView群を責務別モジュールへ分離する。

既存URLと主要なimport/patchパスは互換レイヤーで維持し、Issue 333の調査結果をdocs/researchに追加する。
