+++ app/community/views/manage.py
@@ -1,26 +1,29 @@
 """管理系ビュー: 集会の作成、編集、承認/拒否、閉鎖/再開."""
 import logging
-from datetime import timedelta
 
-import requests
-from django.conf import settings
+import requests  # noqa: F401 - 既存テストの patch パス互換用
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
-from django.core.cache import cache
-from django.core.mail import send_mail
 from django.db import DataError
 from django.shortcuts import get_object_or_404, redirect
-from django.template.loader import render_to_string
-from django.urls import reverse, reverse_lazy
-from django.utils import timezone
+from django.urls import reverse_lazy
 from django.views import View
 from django.views.generic import UpdateView, CreateView, ListView
 
 from event.community_cleanup import cleanup_community_future_data
 
+from ..forms_processor import (
+    approve_community_registration,
+    cleanup_closed_community,
+    close_community_and_cleanup,
+    create_owner_membership,
+    notify_new_community_registration,
+    refresh_calendar_entry_and_event_cache,
+    reject_community_registration,
+)
 from ..forms import CommunitySearchForm, CommunityUpdateForm, CommunityCreateForm
 from ..libs import get_join_type
-from ..models import Community, CommunityMember, WEEKDAY_CHOICES
+from ..models import Community, WEEKDAY_CHOICES
 
 logger = logging.getLogger(__name__)
 
@@ -55,13 +58,7 @@ class CommunityUpdateView(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
     def form_valid(self, form):
         try:
             response = super().form_valid(form)
-            calendar_entry = getattr(self.object, 'calendar_entry', None)
-            if calendar_entry:
-                calendar_entry.save()
-            # カレンダーエントリーに関連するイベントのキャッシュを削除
-            for event in self.object.events.all():
-                cache_key = f'calendar_entry_url_{event.id}'
-                cache.delete(cache_key)
+            refresh_calendar_entry_and_event_cache(self.object)
             messages.success(self.request, '集会情報を更新しました。')
             return response
         except DataError:
@@ -81,25 +78,8 @@ class CommunityCreateView(LoginRequiredMixin, CreateView):
     def form_valid(self, form):
         response = super().form_valid(form)
 
-        # オーナーとしてCommunityMemberを作成
-        CommunityMember.objects.create(
-            community=self.object,
-            user=self.request.user,
-            role=CommunityMember.Role.OWNER
-        )
-
-        # Discord通知を送信
-        if settings.DISCORD_WEBHOOK_URL:
-            waiting_list_url = self.request.build_absolute_uri(reverse('community:waiting_list'))
-            discord_message = {
-                "content": f"**【新規集会登録】** {self.object.name}\n"
-                           f"承認ページ: {waiting_list_url}"
-            }
-            discord_timeout_seconds = 10
-            try:
-                requests.post(settings.DISCORD_WEBHOOK_URL, json=discord_message, timeout=discord_timeout_seconds)
-            except Exception as e:
-                logger.warning(f'Discord通知送信失敗: {e}')
+        create_owner_membership(self.object, self.request.user)
+        notify_new_community_registration(self.object, self.request)
 
         messages.success(self.request, '集会が登録されました。承認後に公開されます。')
         return response
@@ -139,36 +119,7 @@ class AcceptView(LoginRequiredMixin, View):
             return redirect('community:waiting_list')
 
         community = get_object_or_404(Community, pk=pk)
-        community.status = 'approved'
-        community.save()
-
-        # 承認通知メールを送信
-        subject = f'{community.name}が承認されました'
-        my_list_url = request.build_absolute_uri(reverse('event:my_list'))
-        context = {
-            'community': community,
-            'my_list_url': my_list_url,
-            'owner_name': community.get_owner().user_name if community.get_owner() else None,
-        }
-
-        # HTMLメールを生成
-        html_message = render_to_string('community/email/accept.html', context)
-
-        owner_email = community.get_owner_email()
-        if owner_email:
-            sent = send_mail(
-                subject=subject,
-                message='',  # プレーンテキストは空文字列を設定
-                from_email=settings.DEFAULT_FROM_EMAIL,
-                recipient_list=[owner_email],
-                html_message=html_message,
-            )
-            if sent:
-                logger.info(f'承認メール送信成功: {community.name} to {owner_email}')
-            else:
-                logger.warning(f'承認メール送信失敗: {community.name} to {owner_email}')
-        else:
-            logger.warning(f'承認メール送信スキップ: {community.name} - オーナーのメールアドレスが見つかりません')
+        approve_community_registration(community, request)
 
         messages.success(request, f'{community.name}を承認しました。')
         return redirect('community:waiting_list')
@@ -181,34 +132,7 @@ class RejectView(LoginRequiredMixin, View):
             return redirect('community:waiting_list')
 
         community = get_object_or_404(Community, pk=pk)
-        community.status = 'rejected'
-        community.save()
-
-        # 非承認メールを送信
-        subject = f'{community.name}が非承認になりました'
-        context = {
-            'community': community,
-            'owner_name': community.get_owner().user_name if community.get_owner() else None,
-        }
-
-        # HTMLメールを生成
-        html_message = render_to_string('community/email/reject.html', context)
-
-        owner_email = community.get_owner_email()
-        if owner_email:
-            sent = send_mail(
-                subject=subject,
-                message='',  # プレーンテキストは空文字列を設定
-                from_email=settings.DEFAULT_FROM_EMAIL,
-                recipient_list=[owner_email],
-                html_message=html_message,
-            )
-            if sent:
-                logger.info(f'非承認メール送信成功: {community.name} to {owner_email}')
-            else:
-                logger.warning(f'非承認メール送信失敗: {community.name} to {owner_email}')
-        else:
-            logger.warning(f'非承認メール送信スキップ: {community.name} - オーナーのメールアドレスが見つかりません')
+        reject_community_registration(community)
 
         messages.success(request, f'{community.name}を非承認にしました。')
         return redirect('community:waiting_list')
@@ -231,28 +155,10 @@ class CloseCommunityView(LoginRequiredMixin, UserPassesTestMixin, View):
             messages.error(request, '権限がありません。')
             return redirect('community:detail', pk=pk)
 
-        # 閉鎖日を設定（今日の日付）
-        today = timezone.now().date()
-        community.end_at = today
-        community.save()
-
-        # 当日開催分は残し、翌日以降の予定を停止対象にする
-        cleanup_from_date = today + timedelta(days=1)
         try:
-            stats = cleanup_community_future_data(
-                community=community,
-                from_date=cleanup_from_date,
-                delete_rules=True,
-                delete_google_events=True,
-                google_window_days=365,
-                google_years=1,
-            )
-
-            logger.info(
-                f'集会「{community.name}」を閉鎖しました。'
-                f'削除イベント数={stats["db_events"]}、'
-                f'削除定期ルール数={stats["rules"]}、'
-                f'削除Googleイベント数={stats["google_events"]}'
+            stats = close_community_and_cleanup(
+                community,
+                cleanup_func=cleanup_community_future_data,
             )
             messages.success(
                 request,
@@ -283,23 +189,11 @@ class AdminCommunityCleanupView(LoginRequiredMixin, UserPassesTestMixin, View):
 
     def post(self, request, pk):
         community = get_object_or_404(Community, pk=pk)
-        today = timezone.now().date()
 
-        # 閉鎖されていない場合は閉鎖状態にする
-        if community.end_at is None:
-            community.end_at = today
-            community.save(update_fields=['end_at'])
-
-        # 当日開催分は残し、翌日以降を停止対象にする
-        cleanup_from_date = today + timedelta(days=1)
         try:
-            stats = cleanup_community_future_data(
-                community=community,
-                from_date=cleanup_from_date,
-                delete_rules=True,
-                delete_google_events=True,
-                google_window_days=365,
-                google_years=1,
+            stats = cleanup_closed_community(
+                community,
+                cleanup_func=cleanup_community_future_data,
             )
 
             messages.success(
