+++ app/event/forms.py
@@ -1,13 +1,8 @@
-import logging
 from datetime import datetime, timedelta
-from io import BytesIO
-from pathlib import Path
 
 from django import forms
 from django.core.exceptions import ValidationError
-from django.core.files.base import ContentFile
 from django.utils import timezone
-from PIL import Image, UnidentifiedImageError
 
 from community.models import WEEKDAY_CHOICES, TAGS
 from .datetime_lock import (
@@ -16,56 +11,9 @@ from .datetime_lock import (
     has_event_detail_start_time_changed,
     is_event_detail_datetime_locked,
 )
-from .models import EventDetail, RecurrenceRule, Event, validate_pdf_file
-from .thumbnail import SLIDE_THUMBNAIL_ASPECT_RATIO_TEXT, crop_to_slide_thumbnail_aspect_ratio
-
-logger = logging.getLogger(__name__)
-
-
-def _validate_thumbnail_image(thumbnail_image):
-    """サムネイル画像を検証し、スライド比率に中央クロップする."""
-    if not thumbnail_image or not hasattr(thumbnail_image, 'read'):
-        return thumbnail_image
-    if getattr(thumbnail_image, 'size', 0) > 10 * 1024 * 1024:
-        raise ValidationError('画像ファイルサイズが10MBを超えています。')
-
-    try:
-        with Image.open(thumbnail_image) as image:
-            cropped_image = crop_to_slide_thumbnail_aspect_ratio(image.convert('RGB'))
-            image_buffer = BytesIO()
-            cropped_image.save(image_buffer, format='JPEG', quality=90, optimize=True)
-    except (UnidentifiedImageError, OSError):
-        raise ValidationError('有効な画像ファイルをアップロードしてください。')
-    finally:
-        try:
-            thumbnail_image.seek(0)
-        except (AttributeError, OSError, ValueError):
-            logger.exception(
-                "サムネイル画像の読み取り位置リセットに失敗しました: name=%s",
-                getattr(thumbnail_image, 'name', None),
-            )
-
-    filename = f"{Path(thumbnail_image.name).stem}.jpg"
-    return ContentFile(image_buffer.getvalue(), name=filename)
-
-
-def _validate_and_sanitize_pdf(slide_file):
-    """PDFファイルのバリデーションとサニタイズ（共通ヘルパー）"""
-    if not slide_file:
-        return slide_file
-    if not slide_file.name.lower().endswith('.pdf'):
-        raise ValidationError('PDFファイルのみアップロード可能です。')
-    if slide_file.size > 30 * 1024 * 1024:  # 30MB
-        raise ValidationError('ファイルサイズが30MBを超えています。')
-    validate_pdf_file(slide_file)
-    try:
-        slide_file.content_type = 'application/pdf'
-    except (AttributeError, TypeError):
-        logger.exception(
-            "PDFアップロードのcontent_type設定に失敗しました: name=%s",
-            getattr(slide_file, 'name', None),
-        )
-    return slide_file
+from .form_validators import validate_and_sanitize_pdf, validate_thumbnail_image
+from .models import EventDetail, RecurrenceRule, Event
+from .thumbnail import SLIDE_THUMBNAIL_ASPECT_RATIO_TEXT
 
 
 class EventSearchForm(forms.Form):
@@ -385,10 +333,10 @@ class EventDetailForm(forms.ModelForm):
         return cleaned_data
     
     def clean_slide_file(self):
-        return _validate_and_sanitize_pdf(self.cleaned_data.get('slide_file'))
+        return validate_and_sanitize_pdf(self.cleaned_data.get('slide_file'))
 
     def clean_thumbnail_image(self):
-        return _validate_thumbnail_image(self.cleaned_data.get('thumbnail_image'))
+        return validate_thumbnail_image(self.cleaned_data.get('thumbnail_image'))
 
     def save(self, commit=True):
         instance = super().save(commit=commit)
@@ -450,10 +398,10 @@ class LTApplicationEditForm(forms.ModelForm):
         self.initial['generate_blog_article'] = not has_article
 
     def clean_slide_file(self):
-        return _validate_and_sanitize_pdf(self.cleaned_data.get('slide_file'))
+        return validate_and_sanitize_pdf(self.cleaned_data.get('slide_file'))
 
     def clean_thumbnail_image(self):
-        return _validate_thumbnail_image(self.cleaned_data.get('thumbnail_image'))
+        return validate_thumbnail_image(self.cleaned_data.get('thumbnail_image'))
 
     def save(self, commit=True):
         instance = super().save(commit=commit)
