+++ app/user_account/view_modules/__init__.py
@@ -1,0 +1,1 @@
+"""Account view modules."""
